+++ app/src/main/java/com/sqz/checklist/ui/main/task/layout/check/CheckDataState.kt
@@ -2,6 +2,6 @@ package com.sqz.checklist.ui.main.task.layout.check
 
 data class CheckDataState(
     val checkTaskAction: Boolean = false,
-    val undoActionId: Int = -1,
+    val undoActionId: Long = -1,
     val undoButtonState: Boolean = false,
 )
